# api/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "FieldViolation": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "code": {
            "type": "string"
          },
          "currentRevision": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "fieldViolations": {
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            },
            "type": "array"
          },
          "instance": {
            "type": "string"
          },
          "remediation": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "status": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "token",
        "description": "Lab static bearer. Auth is stubbed open in API-001; SEC-001 adds bearer + basic.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST adapter for the shared LabMail capability registry. Generated from internal/capabilities. Do not edit by hand.",
    "title": "LabMail Management API",
    "version": "v1",
    "x-generated-by": "internal/control/rest.RenderOpenAPI; DO NOT EDIT."
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/events/stream": {
      "get": {
        "description": "REST SSE inbox events. MCP uses subscriptions/listen URI-only notify.",
        "operationId": "get_v1_events_stream",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Events stream"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "application/problem+json domain error"
          }
        },
        "summary": "Events stream",
        "tags": [
          "events"
        ],
        "x-capability-id": "events.stream",
        "x-idempotent": true,
        "x-mutating": false,
        "x-required-scopes": [
          "mail.read"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Management listener (default address :1080)",
      "url": "/"
    }
  ],
...
}